+++ bevy_nannou_render/Cargo.toml
@@ -10,4 +10,3 @@ nannou_core = { path = "../nannou_core" }
 nannou_mesh = { path = "../nannou_mesh" }
 rusttype = { version = "0.8", features = ["gpu_cache"] }
 lyon = "0.17"
-bevy_mod_debugdump = "0.9.0"
